Add Mirror<T> and Canonical<T> to share values between threads.
It's annoying to use channels to share values between threads; in order
to get the most up to date value, we need to try_recv() the receiver
until all updates have been processed. This is slow. And if we don't
try_recv(), the receiver's message queue may fill up with values waiting
to be received, which wastes memory. So create a (Mirror, Canonical)
abstraction, which makes it easy to set a value which can be read on
another thread.

The Canonical<T> resides on on thread, and values set() into it can be
read with a Mirror<T>::get() on another thread.

I use interior mutability to share between threads.

Using this should be faster and use less memory for updating the playback
position, buffered ranges, and duration in the example player.